We begin be sketching out our requirements.
In this case, we know that we need to have holiday flights and cargo
flights, so we have created classes for both of them. It would be nice
to treat them together, so we have a Flight role.

I chose to use a role because it defines an abstraction - not an
implementation. Of course, it can have methods, but we can never
instantiate a role. This is perfect for us, and as you'll see later
is great for specifying requirements on what a Flight is.