The first problem is that FlightFactory knows to much.
Our factory doesn't need to know about what creates a specific type
of flight, this is the responsibility of each Flight. So we move
this logic to each Flight, and check in the factory with the new
'understands' class method.

All flights must specify how to understand data, so we require this
method be implemented in our Flight role.