We can be even clearer if we use Module::PluginFinder.
This module abstracts away the idiom of searching for a class given
some predicate. By using this our code is now more readable (it
reads as clearly as it behaves).

It's slightly odd that we have to pass $data twice, but reading the
documentation shows that the first argument is the search key, and the
second are constructor arguments.